# Conflicts:
#	ace/.idea/misc.xml

+++ ace/ace-live/ace-live-service/src/main/java/com/huacainfo/ace/live/service/impl/LiveServiceImpl.java
@@ -3,22 +3,21 @@ package com.huacainfo.ace.live.service.impl;
 
 import java.util.*;
 
-import com.huacainfo.ace.common.tools.DateUtil;
-import com.huacainfo.ace.common.tools.StringUtils;
+import com.huacainfo.ace.common.plugins.wechat.util.Util;
+import com.huacainfo.ace.common.tools.*;
+import com.huacainfo.ace.live.dao.LiveLogDao;
 import com.huacainfo.ace.live.dao.LiveRptDao;
-import com.huacainfo.ace.live.vo.LiveRptQVo;
+import com.huacainfo.ace.live.model.LiveLog;
 import com.huacainfo.ace.portal.model.Users;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.userdetails.User;
 import org.springframework.stereotype.Service;
 
 import com.huacainfo.ace.common.model.UserProp;
 import com.huacainfo.ace.common.result.MessageResponse;
 import com.huacainfo.ace.common.result.PageResult;
 import com.huacainfo.ace.common.result.SingleResult;
-import com.huacainfo.ace.common.tools.CommonUtils;
 import com.huacainfo.ace.live.dao.LiveDao;
 import com.huacainfo.ace.live.model.Live;
 import com.huacainfo.ace.portal.service.DataBaseLogService;
@@ -41,6 +40,8 @@ public class LiveServiceImpl implements LiveService {
     private LiveDao liveDao;
     @Autowired
     private DataBaseLogService dataBaseLogService;
+    @Autowired
+    private LiveLogDao liveLogDao;
 
     /**
      * @throws
@@ -58,18 +59,7 @@ public class LiveServiceImpl implements LiveService {
     @Override
     public PageResult<LiveVo> findLiveList(LiveQVo condition, int start, int limit, String orderBy) throws Exception {
 
-        List<LiveVo> list = this.liveDao.findList(condition,
-                start, start + limit, orderBy);
-        int reportCount;
-        LiveRptQVo rptQVo = new LiveRptQVo();
-        for (LiveVo vo : list) {
-            rptQVo.setRid(vo.getId());
-            rptQVo.setStatus("2");//只统计已发布的报道
-            reportCount = LiveRptDao.findCount(rptQVo);
-            vo.setReportCount(reportCount);
-        }
-
-
+        List<LiveVo> list = this.liveDao.findList(condition, start,  limit, orderBy);
         PageResult<LiveVo> rst = new PageResult<LiveVo>();
         rst.setRows(list);
         if (start <= 1) {
@@ -93,9 +83,8 @@ public class LiveServiceImpl implements LiveService {
     @Override
     public MessageResponse insertLive(Live o, UserProp userProp)
             throws Exception {
-        o.setId(UUID.randomUUID().toString());
-        if (CommonUtils.isBlank(o.getId())) {
-            return new MessageResponse(1, "主键不能为空！");
+        if(CommonUtils.isBlank(o.getId())){
+            o.setId(GUIDUtil.getGUID());
         }
         if (CommonUtils.isBlank(o.getName())) {
             return new MessageResponse(1, "名称不能为空！");
@@ -103,36 +92,17 @@ public class LiveServiceImpl implements LiveService {
         if (CommonUtils.isBlank(o.getCategory())) {
             return new MessageResponse(1, "直播类型不能为空！");
         }
-
         if (CommonUtils.isBlank(o.getStartTime())) {
             return new MessageResponse(1, "开始时间不能为空！");
         }
         if (o.getStartTime().before(DateUtil.getNowDate())) {
             return new MessageResponse(1, "开始时间不能在系统时间之前");
         }
-        if (o.getStartTime().after(o.getEndTime())) {
-            return new MessageResponse(1, "开始时间结束时间之后");
-        }
         if (CommonUtils.isBlank(o.getRemark())) {
             return new MessageResponse(1, "摘要不能为空！");
         }
         if (CommonUtils.isBlank(o.getContent())) {
-            return new MessageResponse(1, "活动介绍不能为空！");
-        }
-        if (CommonUtils.isBlank(o.getNop())) {
-            return new MessageResponse(1, "参与人数不能为空！");
-        }
-        if (CommonUtils.isBlank(o.getPop())) {
-            return new MessageResponse(1, "点赞数不能为空！");
-        }
-        if (CommonUtils.isBlank(o.getAddr())) {
-            return new MessageResponse(1, "地点不能为空！");
-        }
-        if (CommonUtils.isBlank(o.getRtmpUrl())) {
-            return new MessageResponse(1, "直播流地址不能为空！");
-        }
-        if (CommonUtils.isBlank(o.getMp4Url())) {
-            return new MessageResponse(1, "回放地址不能为空！");
+            return new MessageResponse(1, "介绍不能为空！");
         }
         if (CommonUtils.isBlank(o.getImageSrc())) {
             return new MessageResponse(1, "封面不能为空！");
@@ -141,15 +111,17 @@ public class LiveServiceImpl implements LiveService {
         if (temp > 0) {
             return new MessageResponse(1, "直播名称重复！");
         }
-        o.setCreateDate(new Date());
+        o.setNop(Long.valueOf(0));
+        o.setPop(Long.valueOf(0));
         o.setStatus("1");
+        o.setCreateDate(new Date());
         o.setCreateUserName(userProp.getName());
         o.setCreateUserId(userProp.getUserId());
         o.setDeptId(userProp.getCorpId());
         this.liveDao.insert(o);
         this.dataBaseLogService.log("添加直播", "直播", "", o.getName(),
                 o.getName(), userProp);
-        return new MessageResponse(0, "添加直播完成！");
+        return new MessageResponse(0, "提交成功！");
     }
 
     /**
@@ -169,6 +141,10 @@ public class LiveServiceImpl implements LiveService {
         if (CommonUtils.isBlank(o.getId())) {
             return new MessageResponse(1, "主键不能为空！");
         }
+        LiveVo obj=this.liveDao.selectByPrimaryKey(o.getId());
+        if(obj.getAuditStatus().equals("1")||obj.getAuditStatus().equals("2")){
+            //return new MessageResponse(1, "已提交审核，不能编辑！");
+        }
         if (CommonUtils.isBlank(o.getName())) {
             return new MessageResponse(1, "名称不能为空！");
         }
@@ -185,21 +161,7 @@ public class LiveServiceImpl implements LiveService {
         if (CommonUtils.isBlank(o.getContent())) {
             return new MessageResponse(1, "活动介绍不能为空！");
         }
-        if (CommonUtils.isBlank(o.getNop())) {
-            return new MessageResponse(1, "参与人数不能为空！");
-        }
-        if (CommonUtils.isBlank(o.getPop())) {
-            return new MessageResponse(1, "点赞数不能为空！");
-        }
-        if (CommonUtils.isBlank(o.getAddr())) {
-            return new MessageResponse(1, "地点不能为空！");
-        }
-        if (CommonUtils.isBlank(o.getRtmpUrl())) {
-            return new MessageResponse(1, "直播流地址不能为空！");
-        }
-        if (CommonUtils.isBlank(o.getMp4Url())) {
-            return new MessageResponse(1, "回放地址不能为空！");
-        }
+
         if (CommonUtils.isBlank(o.getImageSrc())) {
             return new MessageResponse(1, "封面不能为空！");
         }
@@ -209,55 +171,7 @@ public class LiveServiceImpl implements LiveService {
         this.liveDao.updateByPrimaryKey(o);
         this.dataBaseLogService.log("变更直播", "直播", "", o.getName(),
                 o.getName(), userProp);
-        return new MessageResponse(0, "变更直播完成！");
-    }
-
-    @Override
-    public MessageResponse updateLiveSelective(Live o, UserProp userProp)
-            throws Exception {
-        if (CommonUtils.isBlank(o.getId())) {
-            return new MessageResponse(1, "主键不能为空！");
-        }
-        if (CommonUtils.isBlank(o.getName())) {
-            return new MessageResponse(1, "名称不能为空！");
-        }
-        if (CommonUtils.isBlank(o.getCategory())) {
-            return new MessageResponse(1, "直播类型不能为空！");
-        }
-        if (CommonUtils.isBlank(o.getStartTime())) {
-            return new MessageResponse(1, "开始时间不能为空！");
-        }
-        if (CommonUtils.isBlank(o.getRemark())) {
-            return new MessageResponse(1, "摘要不能为空！");
-        }
-        if (CommonUtils.isBlank(o.getContent())) {
-            return new MessageResponse(1, "活动介绍不能为空！");
-        }
-        if (CommonUtils.isBlank(o.getNop())) {
-            return new MessageResponse(1, "参与人数不能为空！");
-        }
-        if (CommonUtils.isBlank(o.getPop())) {
-            return new MessageResponse(1, "点赞数不能为空！");
-        }
-        if (CommonUtils.isBlank(o.getAddr())) {
-            return new MessageResponse(1, "地点不能为空！");
-        }
-        if (CommonUtils.isBlank(o.getRtmpUrl())) {
-            return new MessageResponse(1, "直播流地址不能为空！");
-        }
-        if (CommonUtils.isBlank(o.getMp4Url())) {
-            return new MessageResponse(1, "回放地址不能为空！");
-        }
-        if (CommonUtils.isBlank(o.getImageSrc())) {
-            return new MessageResponse(1, "封面不能为空！");
-        }
-        o.setLastModifyDate(new Date());
-        o.setLastModifyUserName(userProp.getName());
-        o.setLastModifyUserId(userProp.getUserId());
-        this.liveDao.updateByPrimaryKeySelective(o);
-        this.dataBaseLogService.log("变更直播", "直播", "", o.getName(),
-                o.getName(), userProp);
-        return new MessageResponse(0, "变更直播完成！");
+        return new MessageResponse(0, "提交成功！");
     }
 
     /**
@@ -289,12 +203,17 @@ public class LiveServiceImpl implements LiveService {
      * @version: 2017-12-28
      */
     @Override
-    public MessageResponse deleteLiveByLiveId(String id,
-                                              UserProp userProp) throws Exception {
+    public MessageResponse deleteLiveByLiveId(String id, UserProp userProp) throws Exception {
+        LiveVo o=this.liveDao.selectByPrimaryKey(id);
+        if(CommonUtils.isBlank(o)){
+            return new MessageResponse(1, "提交的主键不存在！");
+        }
+        if(o.getAuditStatus().equals("1")||o.getAuditStatus().equals("2")){
+            return new MessageResponse(1, "已提交审核，不能删除！");
+        }
         this.liveDao.deleteByPrimaryKey(id);
-        this.dataBaseLogService.log("删除直播", "直播", String.valueOf(id),
-                String.valueOf(id), "直播", userProp);
-        return new MessageResponse(0, "直播删除完成！");
+        this.dataBaseLogService.log("删除直播", "直播", String.valueOf(id), String.valueOf(id), "直播", userProp);
+        return new MessageResponse(0, "删除成功！");
     }
 
     @Override
@@ -312,8 +231,6 @@ public class LiveServiceImpl implements LiveService {
         userProp.setCorpId(users.getDepartmentId());
         userProp.setAreaCode(users.getAreaCode());
         userProp.setOpenId(openid);
-        //
-        live.setAreaCode(users.getAreaCode());
         live.setDeptId(users.getDepartmentId());
 
         return insertLive(live, userProp);
@@ -337,6 +254,64 @@ public class LiveServiceImpl implements LiveService {
         response.setOther(data);
         return response;
     }
-
-
+    @Override
+    public  MessageResponse updateLiveAuditByLiveId(String id,String rst,String text, UserProp userProp) throws Exception{
+        if (StringUtils.isEmpty(id)) {
+            return new MessageResponse(1, "直播ID不能为空!");
+        }
+        if (StringUtils.isEmpty(rst)) {
+            return new MessageResponse(1, "审核结果不能为空!");
+        }
+        if (StringUtils.isEmpty(text)) {
+            return new MessageResponse(1, "审核说明不能为空!");
+        }
+        LiveVo liveVo =liveDao.selectByPrimaryKey(id);
+        if(CommonUtils.isBlank(liveVo)){
+            return new MessageResponse(1, "没有获取到直播数据!");
+        }
+        String logId=GUIDUtil.getGUID();
+        LiveLog log=new LiveLog();
+        log.setId(logId);
+        log.setPid(id);
+        log.setAuditor(userProp.getName());
+        log.setRst(rst);
+        log.setStatement(text);
+        log.setCreateDate(new Date());
+        this.liveLogDao.insert(log);
+        this.liveDao.updateAudit(id,rst,logId);
+        if("3".equals(rst)){
+            Map<String ,Object> map=new HashMap<String,Object>();
+            map.put("userId",liveVo.getCreateUserId());
+            map.put("bisType","7");
+            map.put("bisId",id);
+            String url="https://zhsh.wuling.gov.cn/society/pointsRecord/addCoin";
+            HttpUtils.httpPost(url,map);
+        }
+        this.dataBaseLogService.log("直播审核", "直播", "",id, id, userProp);
+        return new MessageResponse(0, "审核成功！");
+    }
+    @Override
+    public  MessageResponse updateStatus(String id,String status, UserProp userProp) throws Exception{
+        if (StringUtils.isEmpty(id)) {
+            return new MessageResponse(1, "直播ID不能为空!");
+        }
+        if (StringUtils.isEmpty(status)) {
+            return new MessageResponse(1, "状态不能为空!");
+        }
+        this.liveDao.updateStatus(id,status);
+        this.dataBaseLogService.log("设置状态", "直播", "",id, id, userProp);
+        return new MessageResponse(0, "设置成功！");
+    }
+    @Override
+    public  MessageResponse updateAuditStatus(String id,String status, UserProp userProp) throws Exception{
+        if (StringUtils.isEmpty(id)) {
+            return new MessageResponse(1, "直播ID不能为空!");
+        }
+        if (StringUtils.isEmpty(status)) {
+            return new MessageResponse(1, "状态不能为空!");
+        }
+        this.liveDao.updateAuditStatus(id,status);
+        this.dataBaseLogService.log("设置状态", "直播", "",id, id, userProp);
+        return new MessageResponse(0, "设置成功！");
+    }
 }
